use correct base class for `Destination`. needs to be `AWSObject`
- the output for destination should be nested under `Property`, using the correct base class achieves this.
- issue cloudtools/troposphere#481
